Broke up Schedule screen into smaller files
